WW-4762 Extracts base abstract class
